Fix ElasticsearchVectorStoreIT FilterExpression with Date type requires the use of epoch milliseconds
